Use wxArrayInt64 for home page

+++ src/mmhomepage.h
@@ -38,7 +38,7 @@ protected:
     wxString title_;
     double grand_total_;
     double grand_gain_lost_;
-    void calculate_stats(std::map<int, std::pair<double, double> > &stockStats);
+    void calculate_stats(std::map<int64, std::pair<double, double> > &stockStats);
 };
 
 class htmlWidgetTop7Categories
@@ -106,7 +106,7 @@ public:
     const wxString displayAccounts(double& tBalance, double& tReconciled, int type);
     ~htmlWidgetAccounts();
 private:
-    std::map<int, std::pair<double, double> > accountStats_;
+    std::map<int64, std::pair<double, double> > accountStats_;
     void get_account_stats();
 };
